% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver_correlation_coefficient.R
\name{Driver_correlation_coefficient}
\alias{Driver_correlation_coefficient}
\title{Driver_correlation_coefficient}
\usage{
Driver_correlation_coefficient("MHY9-EIF3D")
}
\arguments{
\item{x}fusion gene}
\item{y}{svc.df used for mapping fusion genes coordinates to sample structural variant }
\item{z}{CNV segementation table used for mapping fusion gene coordinates to sample segmentation profile}
}
}
\value{
Reports correlation coefficient of structural variation and copy number variation for input fusion gene.
}
\description{
It predicts and plots the correlation coefficient between mapping structural variant and copy number variation profile of sample for a input fusion gene.
}
\examples{
Driver_correlation_coefficient("MYH9-EIF3D")
}
