% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver_domain.R
\name{Driver_domain}
\alias{Driver_domain}
\title{Driver domain}
\usage{
Driver_domain(x)
}
\arguments{
\item{x}{Driver fusion gene}
}
\value{
plot of domain
}
\description{
plotting driver fusion gene domain in cancer
}
