% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver_fusion_result.R
\name{Driver_fusion_result}
\alias{Driver_fusion_result}
\title{Title Driver_fusion_result}
\usage{
Driver_fusion_result("MYH9-EIF3D")
}
\arguments{
\item{x}{ fusion gene}
\item{y}{svc.df used for mapping fusion genes coordinates to structural variants}
\item{z}{CNV segementation table used for mapping fusion gene coordinates to there segmentation profile}
}
\value{
Predicts whether input fusion gene in our datasets is "Driver"" or "Passenger".
}
\description{
Predicts whether input fusion gene in our datasets is "Driver" or "Passenger"" based on mapped the fusion gene to sample structural variant and copy number variation profile.
}
}
\examples{
Driver_fusion_result("MYH9-EIF3D")
}

