% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver_obj.R
\name{Driver_obj}
\alias{Driver_obj}
\title{Driver obj}
\usage{
Driver_obj(x)
}
\arguments{
\item{x}{Driver fusion gene}
}
\value{
plot of domain
}
\description{
plotting driver fusion gene domain in cancer
}
