% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver_plot.R
\name{Driver_plot}
\alias{Driver_plot}
\title{Driver_plot}
\usage{
Driver_plot("MYH9-EIF3D")
}
\arguments{
\item{x}{fusion gene}
}
\value{
Plots fusion gene coordinates with their respective matching structural variant and copy number profile of user dataset.
}
\description{
 It plots the copy number variation profile and structural variants that are mapping to genomic coordinates of input fusion genes.
}
\examples{
Driver_result("MYH9-EIF3D")
}
