% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver_fusion_result.R
\name{Driver_result}
\alias{Driver_result}
\title{Driver_result}
\usage{
Driver_result("MYH9-EIF3D")
}
\arguments{
\item{x}{fusion gene}
\item{y}{svc.df used for mapping fusion genes coordinates to sample structural variant}
\item{z}{CNV segementation table used for mapping fusion gene coordinates to sample segmentation profile}
}
\value{
Predict whether input fusion gene is "Driver" or "Passenger"."
}
\description{
It classifies input fusion gene into “Driver” or “Passenger” fusion gene based on its mapping with sample structural variant and copy number variation.
}
\examples{
Driver_result("MYH9-EIF3D")
}
