% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{input}
\alias{input}
\title{Preparing input files}
\usage{
results_svc <- svc.break.annot(svc, svc.seg.size = 200000, genome.v="hg19",upstr = 100000, verbose=FALSE)
}
\arguments{
\item{x}{svc.df used for mapping fusion genes coordinates to structural variants}
}
\value{
This function incorporates structural variant with gene annotation.
}
\description{
This function creates localization of breakpoints with respect to known genes. The annotation identifies small segmental variants overlapping with genes.
}
x = read.csv("path")
svc = input_svc(x)
results_svc <- svc.break.annot(svc, svc.seg.size = 200000, genome.v="hg19",upstr = 100000, verbose=FALSE)
}
