% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cnv.R
\name{input_cnv}
\alias{input_cnv}
\title{Preparing input files}
\usage{
input_cnv(x)
}
\arguments{
\item{x}{CNV segementation table used for mapping fusion gene coordinates to there segmentation profile}
}
\value{
Preparing input copy number variation file of our sample data form mapping input fusion gene.
}
\description{
It creates input copy number variation file for a user sample data that is used for mapping fusion gene to find out out whether it is “Driver” or “Passenger” fusion gene.


}
\examples{
x = read.csv("path")
cnv = input_cnv(x)
}
