% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_svc.R
\name{input_svc}
\alias{input_svc}
\title{Preparing input files}
\usage{
input_svc(x)
}
\arguments{
\item{x}{svc.df used for mapping fusion genes coordinates to structural variants}
}
\value{
Preparing structural variant file of our sample data form mapping input fusion gene.
}
\description{
It creates input structural varaint file for a user sample data that is used for mapping fusion gene to find out whether it is “Driver” or “Passenger” fusion gene.
}
\examples{
x = read.csv("path")
svc = input_svc(x)
}
