% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver_plot.R
\name{sv.model.view}
\alias{sv.model.view}
\title{Driver_plot}
\usage{
sv.model.view(
  svc,
  cnv,
  chr,
  start,
  stop,
  sampleids = NULL,
  cnvlim = c(-2, 2),
  addlegend = "both",
  cex.legend = 1,
  interval = NULL,
  addtext = NULL,
  cex.text = 0.8,
  plot = TRUE,
  summary = TRUE,
  ...
)
}
\arguments{
\item{x}{Driver fusion gene}
}
\value{
plot of domain
}
\description{
plotting driver fusion gene domain in cancer
}
