% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver_correlation_coefficient.R
\name{Driver_correlation_coefficient}
\alias{Driver_correlation_coefficient}
\title{Driver_correlation_coefficient}
\usage{
Driver_correlation_coefficient(x)
}
\arguments{
\item{x}{fusion gene}
\item{y}{svc.df used for mapping fusion genes coordinates to structural variants}
\item{z}{CNV segementation table used for mapping fusion gene coordinates to there segmentation profile}
}
\value{
It reports the correlation coeffcient of structural variant and copy number varaition for given input fusion gene.
}
\description{
It plots the correlation coefficient between mapping structural variant and copy number variation profile for a given input fusion gene.
}
\examples{
Driver_correlation_coefficient("MYH9-EIF3D")
}

