% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver_domain.R
\name{Driver_domain}
\alias{Driver_domain}
\title{Driver domain}
\usage{
Driver_domain("MYH9-EIF3D")
}
\arguments{
\item{x}{fusion gene}
}
\value{
Plotting fusion gene domain
}
\description{
Provides genomic feature annotation tools for driver fusion gene. Exploring domain level landscape of fusion gene is important to identify gene role in cancer progresssion.
}
}
\examples{
Driver_domain("MYH9-EIF3D")
}

