% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver_obj.R
\name{Driver_obj}
\alias{Driver_obj}
\title{Driver obj}
\usage{
Driver_obj("MYH9-EIF3D")
}
\arguments{
\item{x}{fusion gene}
\item{y}{svc.df used for mapping fusion genes coordinates to structural variants}
\item{z}{CNV segementation table used for mapping fusion gene coordinates to there segmentation profile}
}
\value{
‘Driver_obj’ creates object of input fusion gene based on their mapping structural variant and copy number variation.
}
\description{
It reports the structural variants and copy number variation profile of input fusion that is responsible for their formation.
}
\examples{
Driver_obj("MYH9-EIF3D")
}
